Send remote outlier leave events over federation
When we're in a room, but another HS isn't, we should still send them
relevant leave events (ie to rescind/kick invites).

Because the other HS isn't in the room we won't be running any federation
sender for them, so we wrap the outlier in a to-device event and send it
that way.